snet: Split paths to own file, add PathInterface
The PathInterface is required to losen the dependency on sciond.

Contributes #3348
