+++ go/lib/infra/messenger/BUILD.bazel
@@ -58,7 +58,6 @@ go_test(
         "//go/lib/infra/messenger/mock_messenger:go_default_library",
         "//go/lib/log:go_default_library",
         "//go/lib/overlay:go_default_library",
-        "//go/lib/sciond:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/snet/mock_snet:go_default_library",
         "//go/lib/spath:go_default_library",
