+++ go/lib/snet/mock_snet/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/overlay:go_default_library",
-        "//go/lib/sciond:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/spath:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
