+++ go/lib/svc/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//go/lib/common:go_default_library",
         "//go/lib/l4:go_default_library",
         "//go/lib/overlay:go_default_library",
-        "//go/lib/sciond:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/spath:go_default_library",
         "//go/lib/svc/internal/ctxconn:go_default_library",
